Rename Worker.ts to StanModelWorker.ts
